Exercícios 06/04

+++ Java/dia0604/Ex04.java
@@ -3,18 +3,18 @@ import java.util.Scanner;
 public class Ex04 {
 		public static void main(String[]  args) {
 			Scanner ler= new Scanner(System.in);
-			int[][] a= new int[2][2],b= new int[2][2],c= new int[2][2];
+			float[][] a= new float[2][2],b= new float[2][2],c= new float[2][2];
 			int v,con;
 			for(int i=0;i<2;i++) {
 				for(int j=0;j<2;j++) {
 				System.out.println("Digite o valor da posição ["+i+"]["+j+"] da matriz 1:");
-				a[i][j]=ler.nextInt();
+				a[i][j]=ler.nextFloat();
 				}
 			}
 			for(int i=0;i<2;i++) {
 				for(int j=0;j<2;j++) {
 				System.out.println("Digite o valor da posição ["+i+"]["+j+"] da matriz 2:");
-				b[i][j]=ler.nextInt();
+				b[i][j]=ler.nextFloat();
 				}
 			}
 			System.out.printf("Escolha 1 das opções:\n1-Somar as Matrizes                       3-Adicionar uma constante as Matrizes\n2-Subtrair a primeira matriz da segunda   4-Imprimir as matrizes\n");
@@ -25,7 +25,7 @@ public class Ex04 {
 				for(int i=0;i<2;i++) {
 					for(int j=0;j<2;j++) {
 					c[i][j]=a[i][j]+b[i][j];
-					System.out.printf("[%d]",c[i][j]);
+					System.out.printf("[%.2f]",c[i][j]);
 					}
 					System.out.printf("\n");
 				}
@@ -35,7 +35,7 @@ public class Ex04 {
 				for(int i=0;i<2;i++) {
 					for(int j=0;j<2;j++) {
 					c[i][j]=a[i][j]-b[i][j];
-					System.out.printf("[%d]",c[i][j]);
+					System.out.printf("[%.2f]",c[i][j]);
 					}
 					System.out.printf("\n");
 				}
@@ -47,7 +47,7 @@ public class Ex04 {
 				for(int i=0;i<2;i++) {
 					for(int j=0;j<2;j++) {
 					a[i][j]=a[i][j]+con;
-					System.out.printf("[%d]",a[i][j]);
+					System.out.printf("[%.2f]",a[i][j]);
 					}
 					System.out.printf("\n");
 				}
@@ -56,7 +56,7 @@ public class Ex04 {
 			for(int i=0;i<2;i++) {
 				for(int j=0;j<2;j++) {
 					b[i][j]=b[i][j]+con;
-					System.out.printf("[%d]",b[i][j]);
+					System.out.printf("[%.2f]",b[i][j]);
 					}
 					System.out.printf("\n");
 			}
@@ -65,14 +65,14 @@ public class Ex04 {
 				System.out.printf("Matriz 1:\n");
 				for(int i=0;i<2;i++) {
 					for(int j=0;j<2;j++) {
-					System.out.printf("[%d]",a[i][j]);
+					System.out.printf("[%.2f]",a[i][j]);
 					}
 					System.out.printf("\n");
 				}
 				System.out.printf("\nMatriz 2:\n");
 			for(int i=0;i<2;i++) {
 				for(int j=0;j<2;j++) {
-					System.out.printf("[%d]",b[i][j]);
+					System.out.printf("[%.2f]",b[i][j]);
 					}
 					System.out.printf("\n");
 			}
